품종관리 모달창 한칸씩 밀리는 현상 수정 원종재고관리, 도입자원재고관리, 시교출고관리의 종자보유량, 입고량, 출고량, 재고량 컬럼에서 소수점이 입력되지 않는 문제 해결(mysql 컬럼도 수정 필요)

+++ src/main/java/com/thekiban/Controller/SampleController.java
@@ -449,13 +449,13 @@ public class SampleController {
         } else if (k.equals("종자번호 (ID)")) {
           sampleOutcome.setSample_outcome_num(obj.getString(k));
         } else if (k.equals("종자 보유량")) {
-          sampleOutcome.setSample_outcome_amount(Integer.parseInt(obj.getString(k)));
+          sampleOutcome.setSample_outcome_amount(Double.parseDouble(obj.getString(k)));
         } else if (k.equals("입고량")) {
-          sampleOutcome.setSample_outcome_in(Integer.parseInt(obj.getString(k)));
+          sampleOutcome.setSample_outcome_in(Double.parseDouble(obj.getString(k)));
         } else if (k.equals("출고량")) {
-          sampleOutcome.setSample_outcome_out(Integer.parseInt(obj.getString(k)));
+          sampleOutcome.setSample_outcome_out(Double.parseDouble(obj.getString(k)));
         } else if (k.equals("재고량")) {
-          sampleOutcome.setSample_outcome_remain(Integer.parseInt(obj.getString(k)));
+          sampleOutcome.setSample_outcome_remain(Double.parseDouble(obj.getString(k)));
         } else if (k.equals("담당자")) {
           sampleOutcome.setSample_outcome_person(obj.getString(k));
         } else if (k.equals("인수자")) {
