+++ src/main/java/com/thekiban/Entity/BasicRemain.java
@@ -14,10 +14,10 @@ public class BasicRemain {
   private int basic_remain_id;
   private String basic_name;
   private String basic_remain_num;
-  private int basic_remain_amount;
-  private int basic_remain_in;
-  private int basic_remain_out;
-  private int basic_remain_re;
+  private Double basic_remain_amount;
+  private Double basic_remain_in;
+  private Double basic_remain_out;
+  private Double basic_remain_re;
   private String basic_remain_person;
   private LocalDate basic_remain_date;
   private String create_date;
