+++ src/main/java/com/thekiban/Entity/IncomeRemain.java
@@ -14,10 +14,10 @@ public class IncomeRemain {
   private int income_remain_id;
   private String income_name;
   private String income_remain_num;
-  private int income_remain_amount;
-  private int income_remain_in;
-  private int income_remain_out;
-  private int income_remain_re;
+  private double income_remain_amount;
+  private double income_remain_in;
+  private double income_remain_out;
+  private double income_remain_re;
   private String income_remain_person;
   private LocalDate income_remain_date;
   private String create_date;
